+++ packages/react-core/src/operations/scatterPlot.js
@@ -1,11 +1,17 @@
 import { aggregate } from './aggregation';
 
 // Filters invalid features and formats  data
-export function scatterPlot({ data, xAxisColumns, yAxisColumns, joinOperation }) {
+export function scatterPlot({
+  data,
+  xAxisColumns,
+  xAxisJoinOperation,
+  yAxisColumns,
+  yAxisJoinOperation
+}) {
   return data.reduce((acc, feature) => {
-    const xValue = aggregate(feature, xAxisColumns, joinOperation);
+    const xValue = aggregate(feature, xAxisColumns, xAxisJoinOperation);
     const xIsValid = xValue !== null && xValue !== undefined;
-    const yValue = aggregate(feature, yAxisColumns, joinOperation);
+    const yValue = aggregate(feature, yAxisColumns, yAxisJoinOperation);
     const yIsValid = yValue !== null && yValue !== undefined;
 
     if (xIsValid && yIsValid) {