+++ packages/react-widgets/src/models/CategoryModel.js
@@ -1,10 +1,11 @@
 import { Methods, executeTask } from '@carto/react-workers';
 
 export const getCategories = async (props) => {
-  const { column, operationColumn, operation, filters, dataSource } = props;
+  const { column, operationColumn, joinOperation, operation, filters, dataSource } = props;
 
   return executeTask(dataSource, Methods.FEATURES_CATEGORY, {
     filters,
+    joinOperation,
     operation,
     column,
     operationColumn: operationColumn || column