+++ packages/react-widgets/src/models/ScatterPlotModel.js
@@ -1,11 +1,20 @@
 import { Methods, executeTask } from '@carto/react-workers';
 
 export const getScatter = async (props) => {
-  const { xAxisColumn, yAxisColumn, filters, dataSource } = props;
+  const {
+    xAxisColumn,
+    xAxisJoinOperation,
+    yAxisColumn,
+    yAxisJoinOperation,
+    filters,
+    dataSource
+  } = props;
 
   return executeTask(dataSource, Methods.FEATURES_SCATTERPLOT, {
     filters,
     xAxisColumn,
-    yAxisColumn
+    xAxisJoinOperation,
+    yAxisColumn,
+    yAxisJoinOperation
   });
 };