+++ packages/react-widgets/src/widgets/CategoryWidget.js
@@ -8,6 +8,7 @@ import { getCategories } from '../models';
 import useSourceFilters from '../hooks/useSourceFilters';
 import { selectAreFeaturesReadyForSource } from '@carto/react-redux/';
 import { useWidgetFilterValues } from '../hooks/useWidgetFilterValues';
+import { columnAggregationOn } from './utils/propTypesFns';
 
 const EMPTY_ARRAY = [];
 
@@ -18,7 +19,8 @@ const EMPTY_ARRAY = [];
  * @param  {string} props.title - Title to show in the widget header.
  * @param  {string} props.dataSource - ID of the data source to get the data from.
  * @param  {string} props.column - Name of the data source's column to get the data from.
- * @param  {string} [props.operationColumn] - Name of the data source's column to operate with. If not defined it will default to the one defined in `column`.
+ * @param  {string | string[]} [props.operationColumn] - Name of the data source's column to operate with. If not defined it will default to the one defined in `column`. If multiples are provided, they will be merged into a single one using joinOperation property.
+ * @param  {AggregationTypes} [props.joinOperation] - Operation applied to aggregate multiple operation columns into a single one.
  * @param  {string} props.operation - Operation to apply to the operationColumn. Must be one of those defined in `AggregationTypes` object.
  * @param  {Function} [props.formatter] - Function to format each value returned.
  * @param  {Object} [props.labels] - Overwrite category labels.
@@ -36,6 +38,7 @@ function CategoryWidget(props) {
     dataSource,
     column,
     operationColumn,
+    joinOperation,
     operation,
     formatter,
     labels,
@@ -68,6 +71,7 @@ function CategoryWidget(props) {
       getCategories({
         column,
         operationColumn,
+        joinOperation,
         operation,
         filters,
         dataSource
@@ -87,6 +91,7 @@ function CategoryWidget(props) {
     id,
     column,
     operationColumn,
+    joinOperation,
     operation,
     filters,
     dataSource,
@@ -144,7 +149,11 @@ CategoryWidget.propTypes = {
   title: PropTypes.string.isRequired,
   dataSource: PropTypes.string.isRequired,
   column: PropTypes.string.isRequired,
-  operationColumn: PropTypes.string,
+  operationColumn: PropTypes.oneOfType([
+    PropTypes.string,
+    PropTypes.arrayOf(PropTypes.string)
+  ]),
+  joinOperation: columnAggregationOn('operationColumn'),
   operation: PropTypes.oneOf(Object.values(AggregationTypes)).isRequired,
   formatter: PropTypes.func,
   labels: PropTypes.object,